CoreSet - A generic data container for molecular profiles and treatment response data

CoreSet - A generic data container for molecular profiles and treatment response data

Details

The CoreSet (cSet) class was developed as a superclass for pSets in the PharmacoGx and RadioGx packages to contain the data generated in screens of cancer sample lines for their genetic profile and sensitivities to therapy (Pharmacological or Radiation). This class is meant to be a superclass which is contained within the PharmacoSet (pSet) and RadioSet (rSet) objects exported by PharmacoGx and RadioGx. The format of the data is similar for both pSets and rSets, allowing much of the code to be abstracted into the CoreSet super-class. However, the models involved with quantifying sampleular response to Pharmacological and Radiation therapy are widely different, and extension of the cSet class allows the packages to apply the correct model for the given data.

Slots

annotation

See Slots section.

molecularProfiles

See Slots section.

sample

See Slots section.

treatment

See Slots section.

treatmentResponse

See Slots section.

perturbation

See Slots section.

curation

See Slots section.

datasetType

See Slots section.

Slots

annotation: A list of annotation data about the CoreSet, including the $name and the session information for how the object was created, detailing the exact versions of R and all the packages used.

molecularProfiles: A list or MultiAssayExperiment containing CoreSet object.

sample: A data.frame containg the annotations for all the samples profiled in the data set, across all molecular data types and treatment response experiments.

treatment: A data.frame containing the annotations for all treatments in the dataset, including the mandatory 'treatmentid' column to uniquely identify each treatment.

treatmentResponse: A list or LongTable containing all the data for the treatment response experiment, including $info, a data.frame containing the experimental info, $raw a 3D array containing raw data, $profiles, a data.frame containing sensitivity profiles statistics, and $n, a data.frame detailing the number of experiments for each sample-drug/radiationInfo pair

perturbation: list containing $n, a data.frame summarizing the available perturbation data. This slot is currently being deprecated.

curation: list containing mappings for treatment, sample and tissue names used in the data set to universal identifiers used between different CoreSet objects.

datasetType: character string of 'sensitivity', 'perturbation', or both detailing what type of data can be found in the CoreSet, for proper processing of the data

See also