A helper method to find the best multithreading configuration for your computer

A helper method to find the best multithreading configuration for your computer

Usage

optimizeCoreGx(sample_data, set = FALSE, report = !set)

Arguments

sample_data

TreatmentResponseExperiment

set

logical(1) Should the function modify your R environment with the predicted optimal settings? This changes the global state of your R session!

report

logical(1) Should a data.frame of results be returned by number of threads and operation be returned? Defaults to !set.

Value

If set=TRUE, modifies data.table threads via setDTthreads(), otherwise displays a message indicating the optimal number of threads. If report=TRUE, also returns a data.frame of the benchmark results.

Examples

# \donttest{
  data(merckLongTable)
  optimizeCoreGx(merckLongTable)
#> Benchmarking assay(sample_data, withDimnames=TRUE)...
#> Running with:
#>   nthread
#> 1       1
#> 2       2
#> Benchmarking assays(sample_data)...
#> Running with:
#>   nthread
#> 1       1
#> 2       2
#> Benchmarking reindex(sample_data)...
#> Running with:
#>   nthread
#> 1       1
#> 2       2
#> Optimal cores for your machine are: 2
#>                                          expression nthread     min_sec
#> 1 { assay(sample_data, 1, withDimnames = TRUE) NA }       1 0.004539329
#> 2 { assay(sample_data, 1, withDimnames = TRUE) NA }       2 0.004487482
#> 3                        { assays(sample_data) NA }       1 0.006191011
#> 4                        { assays(sample_data) NA }       2 0.006052130
#> 5                       { reindex(sample_data) NA }       1 0.007117578
#> 6                       { reindex(sample_data) NA }       2 0.007086621
#>    median_sec total_sec mem_alloc_mb  itr/sec   gc/sec
#> 1 0.004599285 0.4786305     3.318416 213.1080 4.178589
#> 2 0.004543717 0.4784025     2.181288 217.3902 4.180581
#> 3 0.006263345 0.4756102     2.738928 157.6922 4.205124
#> 4 0.006126560 0.4743287     2.738928 162.3347 4.216485
#> 5 0.007190354 0.4880153     1.826312 137.2908 2.049116
#> 6 0.007153170 0.4879824     1.537576 139.3493 2.049254
# }